+++ src/features/tracking-main/lib/calulatePercentage.ts
